+++ src/main/java/tamanna/challange/interviews/service/Interviewer/IInterviewerService.java
@@ -11,11 +11,11 @@ public interface IInterviewerService {
 
     List<Interviewer> getAll();
 
-    Optional<Interviewer> getInterviewerById(Integer interviewerId);
+    Optional<Interviewer> getInterviewerById(Long interviewerId);
 
-    void setInterviewerAvailableDates(Integer interviewerId, List<LocalDateTime> availableDates);
+    void setInterviewerAvailableDates(Long interviewerId, List<LocalDateTime> availableDates);
 
     void insertInterviewer(Interviewer interviewers);
 
-    List<AvailableInterviewDates> getAvailableDatesByInterviewer(Integer interviewerId);
+    List<AvailableInterviewDates> getAvailableDatesByInterviewer(Long interviewerId);
 }
